NULL, site_t 修正

+++ 03-2DTransform/Triangle.cpp
@@ -52,8 +52,8 @@ void Triangle::Draw(Renderer& renderer)
 	setupTransform(renderer);
 
 	auto pDeviceContext = renderer.GetDeviceContext();
-	size_t strides[1] = { sizeof(Vertex) };
-	size_t offsets[1] = { 0 };
+	uint32_t strides[1] = { sizeof(Vertex) };
+	uint32_t offsets[1] = { 0 };
 	pDeviceContext->IASetVertexBuffers(0, 1, &VertexBuffer, strides, offsets);
 	pDeviceContext->Draw(VERTEX_NUM, 0);
 }
